Update getJob test to improve readability
The test to assert if the api request is being made on the endpoint is
currently re-purposing the api request function so that the endpoint is
passed to the callback rather than the result.

Assertion can be moved to within the api request itself to make the test
more readable and relevant to what is being tested.